Update ansible with pre-deploy checks.
Also re-orders tasks so that if you just run the playbook at go it
builds both apps, and then moves on to deploying only if both build.
This fixes the possibility that one app builds and deploys, and the
other fails to build, potentially causing a mismatch on the production
systems if there are dependencies between versions.