Refactor configuration and deployment for both apps
Correctly use runtime.exs and it's modern syntax for getting ENV vars.
Also reorganize some of the configuration between environments and
minimize the amount of configuration taking at runtime.

Switch to modern phoenix-generated Dockerfile for building a container
with the application ready to run, and then modify our ansible
deployment technique to extract the application files from a non-running
instance of the built container and copy them to the non-docker
production server.
Note that this allows us to now easily deploy the app as a docker
container in addition to the current dd-app-server method.

Changes the tmp/ directory to _tmp/ to match with _build.
Note that this removes the .gitignores for tmp/ in order to make room
for potentially using that directory in the future, so you may need to
manually remove that folder from existing copies of the repo.